requirements.txt
aiosqlite==0.20.0
annotated-types==0.7.0
anthropic==0.50.0
anyio==4.9.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
executing==2.2.0
fastapi
seaborn
filelock==3.18.0
filetype==1.2.0
fonttools==4.57.0
fsspec==2025.3.2
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc0
google-api-python-client==2.168.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-genai==1.12.1
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
groq==0.23.1
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hypothesis==6.131.9
idna==3.10
importlib_resources==6.5.2
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.7
langchain-anthropic==0.2.1
langchain-cerebras==0.3.0
langchain-core==0.3.56
langchain-google-genai==2.0.10
langchain-groq==0.2.1
langchain-openai==0.2.1
langgraph==0.2.28
langgraph-checkpoint==1.0.12
langgraph-checkpoint-sqlite==1.0.4
langsmith==0.3.38
MarkupSafe==3.0.2
matplotlib==3.9.4
milvus-lite==2.4.12
mistralai==1.7.0
mpmath==1.3.0
msgpack==1.1.0
networkx==3.2.1
numpy==1.26.4
ollama==0.4.7
openai==1.76.0
orjson==3.10.16
packaging==24.2
pandas==2.2.3
pillow==11.2.1
pluggy==1.5.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.10.6
pydantic_core==2.27.2
pymilvus==2.5.8
pymongo==4.10.1
pyparsing==3.2.3
pytest==7.4.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.27.0
sympy==1.13.1
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tomli==2.2.1
torch==2.6.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
varname==0.13.5
websockets==15.0.1
zipp==3.21.0
zstandard==0.23.0
